# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateUserRequest:
    properties:
      age:
        type: integer
      createTime:
        description: 创建时间
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      updateTime:
        description: 更新时间
        type: string
    required:
    - email
    - name
    type: object
...
  models.UpdateUserRequest:
    properties:
      age:
        type: integer
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 这是一个基于Go和Gin框架的后端API服务
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Backend API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: 获取系统中所有用户的列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              count:
                type: integer
              data:
                items:
                  type: object
                type: array
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 获取所有用户
      tags:
      - users
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 创建用户
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 删除用户
      tags:
      - users
    get:
      consumes:
      - application/json
      description: 根据用户ID获取用户详细信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 根据ID获取用户
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 根据ID更新用户信息
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 更新用户
      tags:
      - users
  /users/page:
    get:
      consumes:
      - application/json
      description: 根据分页参数获取用户列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: desc
        description: 排序方式
        in: query
        name: order
        type: string
      - default: create_time
        description: 排序字段
        in: query
        name: order_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  type: object
                type: array
              pagination:
                type: object
              success:
                type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              message:
                type: string
              success:
                type: boolean
            type: object
      summary: 分页获取用户列表
      tags:
      - users
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"